UPSTREAM: thermal: exynos: Use the new thermal trend type for quick cooling action.
This patch uses the quick thermal cooling trend type macros. This is needed
as exynos5 and other thermal sensors now supports only interrupt method for
thresold temperature check.

Acked-by: Kukjin Kim <kgene.kim@samsung.com>
Signed-off-by: Amit Daniel Kachhap <amit.daniel@samsung.com>
Signed-off-by: Zhang Rui <rui.zhang@intel.com>
(cherry picked from linux-next commit
	ce760ed3f440cfa51785c64dad9b4bd87673bb11)

BUG=chrome-os-partner:18622
TEST=Build and boot on peach-pit.  When thermal throttling is enabled,
observe that the frequency is immediately set to the maximum (or
minimum) cooling state when a threshold is crossed.

Change-Id: Ice76ae06665337ce56c87a162a3cf758e8927e5e
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/58055