CHROMIUM: charger-manager: Pass IRQ to wakeup check function
Since charger-manager knows the IRQ number it is looking for in the
wake-up check (it gets it from the RTC class), pass it to the platform
function so that this does not need to be hard-coded.

Add a function to exynos which return true if a particular GPIO is known
to be the sole reason for the wakeup event.

BUG=chrome-os-partner:10617
TEST=build and boot to kernel prompt on snow
Turn on debugging in setup-charger. Try waking up on snow using the
keyboard, the lid and a timer ('rtcwake -s 35 -d /dev/rtc1 -m mem').
See that only the timer causes this function to return true.

Note I did not test simultaneous wakeup. This could perhaps be done by
introducing a long delay in the function, to give me time to cause another
type of wakeup event. However, I did see an EC wakeup event at one point
and this caused the function to return false as required.

Change-Id: I96eb0387e36c097f0df0256c6be4879b3180d259
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43022