drm/i915/ctx: Remove bad invariant
It's not that the assertion is incorrect, but rather that we can call
do_destroy early in loading, and we will falsely BUG().

Since contexts have been in for a while now, and in the internal APIs
are pretty stable, it should be fairly safe to remove this.

v2: Remove unused dev_priv, and dev

BUG=chrome-os-partner:22147
TEST=Sanity Test on Link(unable to repro)

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit f73f760725636b9d0c3786273e185b053516d1eb)
Signed-off-by: David Sodman <dsodman@chromium.org>

Change-Id: I54acf1b0fd3f7c756cfeb8f33a664cc8a30a0b64
Reviewed-on: https://chromium-review.googlesource.com/181023
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: David Sodman <dsodman@chromium.org>
Tested-by: David Sodman <dsodman@chromium.org>