CHROMIUM: coredump: sanitize the setting of signal->group_exit_code
Now that the coredumping process can be SIGKILL'ed, the setting of
->group_exit_code in do_coredump() can race with complete_signal()
and SIGKILL or 0x80 can be "lost", or wait(status) can report
status == SIGKILL | 0x80.

But the main problem is that it is not clear to me what should we
do if binfmt->core_dump() succeeds but SIGKILL was sent, that is
why this patch comes as a separate change.

This patch adds 0x80 if ->core_dump() succeeds and the process was
not killed. But perhaps we can (should?) re-set ->group_exit_code
changed by SIGKILL back to "siginfo->si_signo |= 0x80" in case when
core_dumped == T.

Under review upstream:

http://thread.gmane.org/gmane.linux.kernel/1442910

BUG=chrome-os-partner:16961
TEST=See "coredump: prevent crash pipe reader from blocking suspend."

Signed-off-by: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Change-Id: I96e81bfea6cb5e21e59f2500f502fe1e9c673b43
[sonnyrao: 3.8: moved changes from fs/exec.c to fs/coredump.c]
Reviewed-on: https://gerrit.chromium.org/gerrit/43580
Reviewed-by: Aaron Durbin <adurbin@chromium.org>