CHROMIUM: drm/exynos: call hdmi/mixer commit during resume
During DPMS on, the encoder calls drm_hdmi_apply which will
in turn call mixer_apply and hdmi commit to set the registers
of mixer and hdmi. But in case of system suspend/resume, the
resume function must set the registers in mixer/hdmi.
This has been taken care in fimd but not in hdmi/mixer. Also,
hpd event must not be called during hdmi suspend to ensure
hdmi resumes properly. Even if hdmi is unplugged, during
suspend state, the resume takes care of this by checking
the gpio status.

Original-Change-Id: I810a7cea59d1b14cc2eca64b3fc91f5b5b58bd3f
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
(cherry picked from peach-alpha commit
	f90fc441d10535ad3914702f509022143556ae4c)

Conflicts:
	drivers/gpu/drm/exynos/exynos_mixer.c

BUG=chrome-os-partner:19760
TEST=Build and boot on snow

Change-Id: I334aea445156e0046637adabe8e38b4a20997145
[abrestic: fixed minor conflict due to vblank resume chnage]
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/55922