CHROMIUM: mfd: initialize EC communication earlier
The eDP bridge loaded by the DRM driver needs to be able to control the
screen regulators.
The FETs controlling the screen backlight and VDD are behind the EC, so
we need the EC MFD to be initialized in an earlier phase to allow the
(virtual) TPS65090 driver to register those FETs before the bridge init.

A similar change was done on the SPI version of the EC driver for Pit.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:21952
TEST=boot on Spring and see the traces from the virtual TPS65090
regulators earlier than the display port initialization.

Change-Id: I62223ea9fa79651a787754d45e05dd26edb7a029
Reviewed-on: https://gerrit.chromium.org/gerrit/64870
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>