UPSTREAM: menuconfig: optionally use pkg-config to detect ncurses libs
When building ncurses with --with-termlib several symbols get moved from
libncurses.so to libtinfo.so. Thus when linking with libncurses.so, one
additionally needs to link with libtinfo.so.

The ncurses pkg-config module will be used to detect the necessary libs for
linking. If not available the old heuristic for detection of the ncurses libs
will be used.

Signed-off-by: Justin Lecher <jlec@gentoo.org>
Tested-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
(cherry picked from commit fc9c6e000f62c2eeda426d4d5f022b3d57729bb0)
Signed-off-by: Mike Frysinger <vapier@chromium.org>

BUG=brillo:1008
TEST=precq passes

Change-Id: I70c1f96d44de0f5f57ec80093e3cfcc940eac2bb
Reviewed-on: https://chromium-review.googlesource.com/274752
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>